- started abstracting anti-bounce-back outlet to work on any side of the domain (not completed yet)  - removed unused functions from obstacle classes
